+++ tests/testthat/test-http.R
@@ -177,3 +177,12 @@ test_that("parse and build are symmetric", {
   round_trip("https://google.com:80/a/b")
   round_trip("https://google.com:80/a/b/")
 })
+
+test_that("rcf2616 returns an ASCII date and undoes changes to the locale", {
+  old <- Sys.getlocale("LC_TIME")
+  defer(Sys.setlocale("LC_TIME", old))
+
+  date <- rfc2616Date(time = as.POSIXct("2024-01-01 01:02:03", tz = "EST"))
+  expect_equal(date, "Mon, 01 Jan 2024 06:02:03 GMT")
+  expect_equal(Sys.getlocale("LC_TIME"), old)
+})