+++ NEWS.md
@@ -1,5 +1,7 @@
 # rsconnect (development version)
 
+* Restore the `LC_TIME` locale after computing an RFC-2616 date. (#1035)
+
 # rsconnect 1.2.0
 
 * Addressed a number of republishing and collaboration issues where the